requirements.txt
scikit-learn
pandas
numpy
streamlit
pyyaml
